Library

Library

Documentation for SplineFEEC.jl's interface.

Bases

SplineFEEC.BSplineTensorProductBasis — Type.

BSplineTensorProductBasis{N,T}(Bases::SArray)

BSpline basis on a tensor-product domain, consisting of a vector of one-dimensional bases defined on each direction of the tensor-product domain.
Where `N` describes the dimension and `T` the boundary condition.

source

Base.adjoint — Method.

adjoint(b::Basis)

Create a basis that consits of the derivative of basis functions in `b`.

source

SplineFEEC.PeriodicBSpline — Method.

PeriodicBSpline(basis_functions, knot_vector, interval, order, N)

Construct a `periodic` BSpline basis on `interval` with `N` points and order `k`.

source

Utilities

SplineFEEC.mass_matrix — Method.

mass_matrix(b::Basis)

Calculate the mass-matrix for the basis `b` with Gauss quadrature.

source

SplineFEEC.stiffness_matrix — Method.

stiffness_matrix(b'::Basis)

Calculate the stiffnes-matrix for the basis `b` with Gauss quadrature. Caution: We directly input b' here as a basis.

source

Operators

Circulant